[Chromium/Mac] Don't send an onclick event from a ctrl-click https://bugs.webkit.org/show_bug.cgi?id=105554
Patch by Alexei Svitkine <asvitkine@chromium.org> on 2012-12-21
Reviewed by Dimitri Glazkov.

Don't send the onclick event on a control click event (which
triggers the context menu on the Mac). This matches the
behaviour of the Apple Mac port, although the Apple port
does this by checking whether the context menu is visible
in WebPage::mouseEvent().

This is a fix for http://crbug.com/146143

Source/WebCore:

Test: fast/events/controlclick-no-onclick.html

* page/EventHandler.cpp:
(WebCore::EventHandler::handleMouseReleaseEvent):

LayoutTests:

* fast/events/controlclick-no-onclick-expected.txt: Added.
* fast/events/controlclick-no-onclick.html: Added.

Identifier: 123865@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138380 268f45cc-cd09-0410-ab3c-d52691b4dbfc