[BlackBerry] Use fatfingers result when checking for misspelled words https://bugs.webkit.org/show_bug.cgi?id=105923
Reviewed by Yong Li.

PR270866
Currently we are using the raw touch point when checking to see if spell check
is required on a word. When tapping close to a boundary, the raw point can be
placed off but fatfingers moves the point onto the word. This causes the caret
to be placed on a word but no spell check request triggered.

Internally reviewed by Mike Fenton and Tiancheng Jiang

* WebKitSupport/TouchEventHandler.cpp:
(BlackBerry::WebKit::TouchEventHandler::handleTouchPoint):


Identifier: 124103@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138619 268f45cc-cd09-0410-ab3c-d52691b4dbfc