https://bugs.webkit.org/show_bug.cgi?id=105475 Nested fixed position elements incorrectly repositioned on scrolling
Reviewed by Simon Fraser.

Source/WebCore:

The fix here is to ensure that the ScrollingTree reflects the same hierarchy as
the GraphicsLayer tree. And from there, we can correctly calculate the offset for
nested objects by keeping track of a cumulative delta that parents have been moved
by.

We already had to fix this same bug in ScrollingStateNode::removeChild(). The
index can be notFound when you are removing a grandchild.
* page/scrolling/ScrollingTreeNode.cpp:
(WebCore::ScrollingTreeNode::removeChild):

parentScrollPositionDidChange() now takes a second parameter that represents the
delta that parent graphicsLayers have been moved by.
* page/scrolling/ScrollingTreeNode.h:
(ScrollingTreeNode):
* page/scrolling/ScrollingTreeScrollingNode.h:
* page/scrolling/mac/ScrollingTreeFixedNode.h:
(ScrollingTreeFixedNode):

attachToStateTree() now handles re-parenting.
* page/scrolling/mac/ScrollingCoordinatorMac.mm:
(WebCore::ScrollingCoordinatorMac::attachToStateTree):

Properly adjust by and compute the new parameter cumulativeDelta. Adjusting by
this will prevent fixed children of fixed objects from moving when they are not
supposed to.
* page/scrolling/mac/ScrollingTreeFixedNode.mm:
(WebCore::ScrollingTreeFixedNode::parentScrollPositionDidChange):

Give parentScrollPositionDidChange() an initial delta of (0,0).
* page/scrolling/mac/ScrollingTreeScrollingNodeMac.mm:
(WebCore::ScrollingTreeScrollingNodeMac::setScrollLayerPosition):

Properly adjust by and compute the new parameter cumulativeDelta. This math is
actually not quite right for sticky yet as the FIXME indicates.
* page/scrolling/mac/ScrollingTreeStickyNode.h:
(ScrollingTreeStickyNode):
* page/scrolling/mac/ScrollingTreeStickyNode.mm:
(WebCore::ScrollingTreeStickyNode::parentScrollPositionDidChange):

Re-named attachToScrollingCoordinator() to attachToScrollingCoordinatorWithParent
to make it clear that the node may already be attached, but it may need to be re-
parented.
* rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::attachToScrollingCoordinatorWithParent):
* rendering/RenderLayerBacking.h:
(RenderLayerBacking):
* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::updateBacking):

Always call attachToScrollingCoordinatorWithParent() in case the fixed node has
been re-parented.
(WebCore::RenderLayerCompositor::registerOrUpdateViewportConstrainedLayer):

LayoutTests:

Unfortunately, the children were not really accurate before, and they are not
really accurate now. It seems like the state tree is not fully up-to-date when the
test is printing it.
* platform/mac/tiled-drawing/fixed/nested-fixed-expected.txt:



Identifier: 124234@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138758 268f45cc-cd09-0410-ab3c-d52691b4dbfc