WebKitCSSTransformValue::customCssText() should always allocate once https://bugs.webkit.org/show_bug.cgi?id=105445
Reviewed by Alexis Menard.

When converting a WebKitCSSTransformValue to a String representation,
it was typically done with 2 memory allocations due to the string size in StringBuilder.

This patch changes the code to use the String Concatenate operations in order
to create the string at once.

The opening parenthesis is also moved in the prefix name in order to maximize
the range copied with memcopy.

* css/WebKitCSSTransformValue.cpp:
(WebCore::transformValueToCssString): New generic function for boxing values.
(WebCore::WebKitCSSTransformValue::customCssText):
(WebCore::WebKitCSSTransformValue::customSerializeResolvingVariables):


Identifier: 124304@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138829 268f45cc-cd09-0410-ab3c-d52691b4dbfc